refactor: extract CardsList, update file names and page layout
